added sorting function
