+++ src/UIkit/TableKit/TableHead/TableHead.tsx
@@ -1,15 +1,25 @@
 
+import { TableHTMLAttributes } from 'react'
 import { TableRow } from '../TableRow/TableRow'
 
-export const TableHead = ({ titleArray }: TableHeadProps) => {
+export function TableHead<T>({ titles, onHeadRowClick, onHeadСellClick, ...props }: TableHeadProps<T>) {
+    const keysOfTitles = Object.keys(titles) as (keyof T)[]
     return (
-        <thead>
-            <TableRow>
-                {titleArray.map(title => <th key={title}>{title}</th>)}
+        <thead >
+            <TableRow onClick={onHeadRowClick ? (e) => onHeadRowClick(e) : undefined} {...props}>
+                {keysOfTitles.map(keyOfTitle =>
+                    <th key={String(keyOfTitle)} onClick={onHeadСellClick ? (e) => onHeadСellClick(e, keyOfTitle) : undefined}
+                    >
+                        {titles[keyOfTitle]}
+                    </th>
+                )
+                }
             </TableRow>
         </thead>
     )
 }
-interface TableHeadProps {
-    titleArray: string[]
+interface TableHeadProps<T> extends TableHTMLAttributes<HTMLTableRowElement> {
+    titles: Record<keyof T, string>
+    onHeadRowClick?: (event: React.MouseEvent<HTMLTableRowElement>) => void
+    onHeadСellClick?: (event: React.MouseEvent<HTMLTableCellElement>, title: keyof T) => void
 }
