+++ src/UIkit/index.tsx
@@ -1,4 +1,6 @@
 export { Button } from './Button/Button'
 export { TextInput } from './TextInput/TextInput'
 export { Title } from './Title/Title'
-export { Table } from './TableKit/Table/Table'
+export { TableBody } from './TableKit/TableBody/TableBody'
+export { TableHead } from './TableKit/TableHead/TableHead'
+export { TableRow } from './TableKit/TableRow/TableRow'
