+++ src/components/index.tsx
@@ -1,0 +1,1 @@
+export { Table } from './Table/Table'
