+++ src/modules/UsersTable/index.tsx
@@ -1,0 +1,1 @@
+export { UsersTable } from './components/UsersTable/UsersTable'
